+++ tests/Unit/BaseTestCase.php
@@ -2,11 +2,18 @@
 
 namespace DDTrace\Tests\Unit;
 
+use DDTrace\Configuration;
 use PHPUnit\Framework;
 
 
 abstract class BaseTestCase extends Framework\TestCase
 {
+    protected function setUp()
+    {
+        Configuration::clear();
+        parent::setUp();
+    }
+
     protected function tearDown()
     {
         \Mockery::close();